requirements.txt
asgiref==3.5.2
attrs==22.1.0
backports.zoneinfo==0.2.1
bcrypt==4.0.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==37.0.4
distro==1.7.0
Django==4.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
docker==6.0.0
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
idna==3.3
jsonschema==3.2.0
packaging==21.3
paramiko==2.11.0
psycopg2==2.9.3
pycparser==2.21
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dotenv==0.20.0
pytz==2022.2.1
PyYAML==5.4.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
texttable==1.6.4
urllib3==1.26.12
websocket-client==0.59.0
